Rename application command files
