+++ commands/hub/help.js
