+++ commands/hub/invite.js
