+++ commands/maps/arenas.js
