+++ commands/maps/battle-royale.js
