+++ commands/maps/control.js
